feat: original script by drdrang
Taken directly from [Dr. Drang](https://github.com/drdrang)'s blog post
at http://www.leancrew.com/all-this/ – archived in the WayBack Machine at
https://web.archive.org/web/20120920012828/http://www.leancrew.com/all-this/

This script is a marvel of elegance and simplicity, and although written for
Python 2, requires only the addition of parentheses on the last line to run
over a decade later on Python 3.12. That's not just a feature, it's a feat.