fix: fatal exception in tidy-md-refs.py
Update regular expressions and code to handle missing reference
link definitions and CommonMark compatibility for indented reference
definitions.

Add example before-and-after Markdown files to tests/ for regression chccks.

When a Markdown input has a reference link that is undefined,
tidy-md-refs.py throws a KeyError exception:

```
% ./tidy-md-refs.py tests/untidy.md
Traceback (most recent call last):
  File "/Users/alexdupuy/Work/reliabot/./tidy-md-refs.py", line 117, in <module>
    sys.exit(main())
  File "/Users/alexdupuy/Work/reliabot/./tidy-md-refs.py", line 111, in main
    if not tidy_file(in_file, out_file):
  File "/Users/alexdupuy/Work/reliabot/./tidy-md-refs.py", line 69, in tidy_file
    tidied = tidy(original)
  File "/Users/alexdupuy/Work/reliabot/./tidy-md-refs.py", line 53, in tidy
    newlabels = [
  File "/Users/alexdupuy/Work/reliabot/./tidy-md-refs.py", line 54, in <listcomp>
    "[%d]: %s" % (i + 1, labels[j]) for (i, j) in enumerate(order)
KeyError: '2'
```

Additionally, it parses reference definitions (labels) to include not
only the following line (as specified by CommonMark) bit also all
lines up to the next non-blank line. This isn't usually a problem,
but if the next non-blank line is also a reference definition,
the errors can cascade.

The exception is easy to fix, the reference definition is much harder.
This commit isn't 100% correct but avoids most problems, and allows
for up to three spaces of indentation for reference defintinos
(as specified by CommonMark). Handling reference titles that extend
to a second following line (or beyond) is *not* supported, and would
be very difficult to implement withoug a full CommonMark parser.